Plumb support for mounting a sandboxfs instance during a build.
This introduces user-facing options to enable the experimental sandboxfs
support and, when enabled, mounts a sandboxfs instance throughout the
build.  The sandboxfs' process handle is passed to the
SandboxActionContextProvider so that the SpawnRunners can later consume it.

Note that this does NOT yet provide sandboxfs support for the builds as the
SpawnRunners are untouched.

RELNOTES: None.
PiperOrigin-RevId: 189678732