Automated rollback of commit 72d28f3efc2842510a34cacd930c0204143f7412.
Fix skylark caching to properly include transitive dependencies when there is a diamond-like dependency in the loaded bzl files.

Also add guards to make sure we're not attempting to cache skylark files that transitively request a dependency that is in error.

*** Reason for rollback ***

Looking for source of non-determinism

*** Original change description ***

Automated rollback of commit 7ba939dfd5df48903929e9c14ebd0449656403e4.

*** Reason for rollback ***

Likely cause for non-determinism in skyframe

*** Original change description ***

Cache SkylarkLookupImportValues in memory so that we don't recompute them multiple times.

PiperOrigin-RevId: 189686604