Add more documentation for rule.outputs
We want to simplify/remove this API, but let's document it in the meantime.

Also fix harmless bug that annoyed me. It used rawTemplate instead of formatStr when there are no placeholders. This would theoretically fail to properly escape % symbols, except that the code path luckily short-circuits later on if there are no placeholders.

RELNOTES:
PiperOrigin-RevId: 189760809