Add a test, verifying that paths within the same repo can be hard-coded
To simplify the transition for existing repositories, we keep the property
that paths to true source files in your own repository may be hard-coded.
Add a test verifying that we keep this property.

Also add a test checking that this property is true, if the same repository
is built as a remote repository. But disable this test, as this property
does not hold yet; we plan to make it true with a change of the execroot
layout.

Change-Id: I0fb961f8ddb8feb98a6309533fc82f2acd9afd0d
PiperOrigin-RevId: 189898773