Remove support for dynamically configured resource filtering
Dynamically configured resource filtering was never turned on, as dynamic
configuration turned out to be too slow.

While we're at it, do a bit of related cleanup (remove the unused flag that
controlled resource filtering in analysis, and switch from a now-two-valued
enum to a boolean to specify whether resources should be filtered in analysis).

RELNOTES: none
PiperOrigin-RevId: 189923588