Update the client's skeleton logging framework to use it for --client_debug.
We are still unable to turn this on to write to files, but there are currently 2 logging systems in use in the client: the inactive one, and the print-to-stderr option triggered by --client_debug. Combine these, so we can use the same logging format for both.

Also combine it with the VerboseLogging functionality - it was not documented anywhere and seems redundant.

RELNOTES: None.
PiperOrigin-RevId: 189979051