Remove current contents of the default global bazelrc.
The file is left as an empty file, so make upgrades easier. Site administrators can add their own content as needed, still.

Fixes #4850.
Closes #4857.

TESTED: Created a Debian VM and tested installing the new version as an upgrade and as the first Bazel, observed that the global bazelrc had the expected content.
RELNOTES: Remove the default content of the global bazelrc.

Change-Id: I5f285f0c59d18179005ffe5bfc35931299635bbb
PiperOrigin-RevId: 190060838