Make ManifestMergerAction able to not take a primary manifest
This is the only difference in action implementation required for the new Skylark Android Data API design. Even if rules have resources or assets, after this API is adopted, they will no longer have to specify a primary manifest. Instead, a dummy manifest will be generated (either to merge with dependency manifests, or for use by IDEs if there are no dependency manifests).

Support this in the ManifestMergerAction by supporting not having a --manifest value passed.

RELNOTES: None
PiperOrigin-RevId: 190095623