Add support for sandboxfs-based sandboxing to macOS.
This only modifies the DarwinSandboxedSpawnRunner for now and leaves behind
the Linux version.

The latter is a bit more complex to do because of the need to deal with tmpfs
and bind-mount interactions.  We expect sandboxfs to have the biggest impact
on macOS anyway, not Linux, so this can wait until this has proven that it
deserves staying for the long-term.

Tested: Manually ran a self-build of Bazel on a version of Bazel built with
these changes and the build passed.

RELNOTES: None.
PiperOrigin-RevId: 190133429