Don't bound the method and field name caches in FuncallExpression.
The number of *unique* Class instances that will be used as keys in these caches over the lifetime of a Bazel server is small-ish, and bounded by the size of the set of unique Classes used as the Java runtime representation of Skylark objects.

The bookkeeping done by LocalCache when the cache's size is bounded is therefore pure CPU overhead.

While we're here, just use the default capacity (16); a smaller initial capacity seems silly considering these caches are live for the duration of the Bazel server.

RELNOTES: None
PiperOrigin-RevId: 190238099