Isolate C++ compile build variables
This is a preparation work to expose Variables instance for all compile actions
to Skylark.

I didn't do linking variables in this cl, because this cl is already too big.
But they're coming shortly in a separate cl.

This is also in line with our goal to make build variables more discoverable and
better document.

RELNOTES: None.
PiperOrigin-RevId: 190591080