Use JUnit assertions instead of Truth assertions.
The format of Truth's failure messages is about to change, which would break these tests.

Alternatively, I could modify these tests to accept both the old and new Truth message formats, but ideally most tests wouldn't rely on the exact message format at all.
(That suggests that an even better approach might be to just throw an exception whose message is a fixed string. That would avoid relying even on JUnit's message format, unlikely as I think that is to change.)

RELNOTES: None.
PiperOrigin-RevId: 190617095