C++: Remove headers from input to linking action.
Headers were made an input to the linking action simply to get an error when we have a src-less cc_library with declared but missing headers. Since there was no compilation action, the headers were not an input to any action and there was no error when the files were missing.

After discussing it with the team, it was decided that this is not needed anymore. Files can be missing as long as they are not consumed.

RELNOTES:none
PiperOrigin-RevId: 190915591