Rename LocalResourceContainer to AndroidResources and remove asset code from it
As part of decoupling Android resources and assets, rename
LocalResourceContainer to AndroidResources and remove asset code from it. Some
general asset and manfiest code still remains and will be dealt with in future
changes.

Remove LocalResourceContainer from the ParsingActionBuilder, since it's always
used to build the ResourceContainer that is subsequently passed in.

RELNOTES: none
PiperOrigin-RevId: 190945260