Remove CcToolchainProvider#getEnvironment() and all the supporting infrastructure.
This was added in unknown commit to provide a different environment to Apple toolchains, then its use removed in unknown commit in favor of getting the environment variables from the CToolchain proto.

I haven't done my research if that's a better approach, but it looks like it (the less hard-coded stuff we have in Java, the better), but worst of all is surely to have *two* such mechanisms.

RELNOTES: None.
PiperOrigin-RevId: 191411878