Remove hashCode int field from Node.
This is needless of complexity. Having counter in DiGraph can provide us with deterministic hashCode for every application run. There is no visible usage of this feature right now. More than, user of the DiGraph class should not rely on the fact the hashCode of the object will always be deterministic. Default Object.hashCode implementation returns deterministic result in cope of single JVM, and this is enough.

RELNOTES:none
PiperOrigin-RevId: 192093877