Made 'file_posix.cc' POSIX compatible
The `d_type` field is not part of the POSIX specification. Added a compile time check to see if we can use it. When not present fallback to a (slightly more expensive) call to 'stat'.

The reason why I need this is because I'm trying to port Bazel to a POSIX compliant platform. Even though my porting effort might fail miserably ;-) I think having a POSIX implementation which is POSIX compliant would be of benefit to the Bazel project.

Besides the POSIX spec I've based the implementation on information I found here:

* https://stackoverflow.com/questions/2197918/cross-platform-way-of-testing-whether-a-file-is-a-directory
* https://stackoverflow.com/questions/23958040/checking-if-a-dir-entry-returned-by-readdir-is-a-directory-link-or-file#answer-29094555
* https://stackoverflow.com/questions/39429803/how-to-list-first-level-directories-only-in-c/39430337#39430337

Closes #4967.

PiperOrigin-RevId: 192102522