Fix handling of relative symlinks within sandboxfs.
If an action expresses a symlink as an input, the target of the symlink does
not necessarily appear as a file to map within the sandbox.  This is a
problem when the target of the symlink is relative because sandboxfs would
expose the link verbatim and the target would be missing later on during
resolution.

To fix this, special-case the handling of symlinks: when trying to expose
them via a sandboxfs mount point, resolve their final target instead of
respecting the original contents.  This loses the fact that the file was a
symlink when running within the sandboxfs sandbox, but is easier to
implement and slightly faster at runtime.  We can reconsider this choice
if this causes problems.

RELNOTES: None.
PiperOrigin-RevId: 192325932