Add CommandLinesAndParamFiles class.
This class is currently unused. In order to keep CL sizes down we introduce this class prior to using it in the spawn runners.

This class can manage an action's list of command lines and param files. For instance, SpawnAction will contain one of these instances (instead of keeping a single command line object and adding param file write actions to the action graph).

At spawn execution time, the spawn runners will use this class to resolve the list of command lines and param files into a master argument list + some number of param files that need to be written. The local spawn runners can simply write the param files using a helper. In the distributed cases the param files are ready-made VirtualActionInputs that can be added to the Spawn's other inputs.

RELNOTES: None
PiperOrigin-RevId: 192439501