Automated rollback of commit 7edc5d49c8e764ab2bdf19e3a9fb58ea30fa708f.
*** Reason for rollback ***

b/77997997

*** Original change description ***

Remove the Blaze strict java deps exemption for Android databinding. Use the JavaBuilder processor exemption list instead of disabling strict deps for the affected rules outright. This narrows the degree to which the exemption is applied, and prevents other strict deps violations from accruing on targets that enable databinding.

PiperOrigin-RevId: 192755946