Resource-only merge action
Create wrapper for merged resources, and methods to make it from parsed
resources.

Currently just reuse the ResourceDependencies class to provide resource
dependencies; we could create an interface and another class, but this class is
good enough (and will be resource-specific once we finish decoupling assets,
resources, and manifests).

Note that, like the previous parse action, merging sometimes has a dependency
on the stamped manifest, so resources and manifests will only be mostly
decoupled.

Also, make the merge action not require an output manifest for cases like this.

RELNOTES: none
PiperOrigin-RevId: 192805891