Resource processing pipeline uses interface instead of ResourceContainer
This allows us to use either the new AndroidResources object and its children (which are fully decoupled from asset processing) or the old ResourceContainer. We'll actually turn on decoupled asset and resource processing in an upcoming change.

We still need to create an asset-only pipeline and handle the binary case.

RELNOTES: none
PiperOrigin-RevId: 193047924