Add an info-key for the repository cache
Add a key to `bazel info` to obtain the location of the repository
cache. As, by default, it is some directory under the bazel output
root, its location is not obvious. Moreover, knowing the location
is useful to have direct access to the downloaded artifacts for
further use (e.g., for prefilling caches on local machines from
a CI system).

Change-Id: I95b56ed424c261e5b5185f02fcb82c5df72ef83a
PiperOrigin-RevId: 193162119