Add helpful functionality to CommandLines.
We're mostly adding convenience methods. This will be used by spawn runners and in Spawn construction, but we break it out now to lessen the size of the final CL.

We also pull out a CommandLineLimits class instead of using an int. This will partially help with readability, but it also likely that we will add a second constraint (max length of a single argument).

RELNOTES: None
PiperOrigin-RevId: 193191097