Wire up decoupled data pipeline in top-level Android targets
The decoupled pipeline is safely hidden behind a flag.

I'm explicitly holding off on the MobileInstall code until next review, so we
sometimes create an ApplicationManifest object even in the new pipeline to pass
data to that code.

I'm not thrilled with the 'just call into the resource processor with different
settings' approach that I've reproduced in ProcessedAndroidData, but I don't
see a better way of doing things, and I think that breaking out the commonly
used settings makes things cleaner.

RELNOTES: none
PiperOrigin-RevId: 193373347