Wire up decoupled resource processing for MobileInstall
This should be the last place the new data processing is needed.

MobileInstall is a bit trickier than normal data processing because it's out of
the way. I'm still using ApplicationManifest to wrap the manifest Artifact in a
lot of it, just because that makes more sense than passing around both
ApplicationManifest for use with old behavior and StampedAndroidManifest for
use with the new. Once we switch to entirely the new behavior, we can remove
ApplicationManifest entirely.

RELNOTES: none
PiperOrigin-RevId: 193392923