Automated rollback of commit 559ffb7dd6578c961e775ba4901917ab8bffcb9d.
*** Reason for rollback ***

Looks like this is using more cpu than I realized. Doesn't seem fixable short-term. This may need to be cherrypicked into the next canary if it shows up in profiling. See bug for details.

*** Original change description ***

Add ValueConstants helper to handle the increasingly complex logic for value-equality-tested constants. Main hurdle is efficiently testing to see if a Collection is a value constant without trying to do work on Collections that can't possibly be value constants.

PiperOrigin-RevId: 193403226