Rename SpawnExecutionPolicy -> SpawnExecutionContext.
This class will be used to tie a Spawn to a SpawnRunner, and isn't really a policy object. It will carry state such as the expanded inputs and expanded command line.

Currently a context can be passed between different SpawnRunners. This will be addressed independently, so a context is tied to a particular spawn runner.

PiperOrigin-RevId: 193501918