Reduce the state kept in OptionProcessor.
Make the list of rc files a local variable as it need not be a class
attribute, and drop the unused rcoptions_ field.

This is a trivial refactoring and the remaining code is still too
confusing.  It'd be worth splitting OptionProcessor in two pieces:
OptionProcessor to exclusively keep the virtual ParseOptions method
and no state, and a new ParsedOptions type to act as the immutable
return value of ParseOptions.  This would decouple all state
mutations.

RELNOTES: None.
PiperOrigin-RevId: 193557347