Fix up all the tests to work with "blaze run --direct_run".
In particular, update run_test so that the little test binary used to check ANSI escape code stripping explicitly marks the beginning of its output and delete a bunch of tests run RunCommandTest that verified the output of the binary. They are not relevant anymore now that the output is not piped through the server.

Also drive-by fix a few linter issues in RunCommand.

RELNOTES: None.
PiperOrigin-RevId: 193655720