Allow repository rules to return a value
...in preparation of the change towards resolved workspaces,
where a source-like rule is supposed to return the information
needed to make it reproducible. At the moment, the value is
ignored (it is just printed back to the user), but it allows
rule authors to prepare for the upcoming change.

RELNOTES: repository rules are no longer restricted to return None.

Change-Id: I1b3bdc7b02a957f00c9846bcb0ee6b7e09343d72
PiperOrigin-RevId: 193664946