Fix asset processing bug
We need to pass primary asset and resource artifacts to processing, not just
resources.

This was introduced as part of my recent migration. In my defense, this migration was needed exactly to avoid situations like this (ResourceContainer.getArtifacts does not just return resource artifacts like you might expect it to).

RELNOTES: none
PiperOrigin-RevId: 193693868