Add the additional options to support manual trimming of feature flags.
There are effectively three different states a flag's value could be in:
1. Value is known to be non-default
2. Value is known to be default
3. Value is unknown (has been trimmed)

In addition to flagValues (which covers the first state), there are now
two additional sets covering the other two states. Neither of these sets
are used when manual trimming is disabled or when the entire set of flags
is known, in which case state 1 is represented by labels in the map,
state 2 is represented by labels not in the map, and state 3 doesn't exist.

This also adds the flag which controls whether manual trimming is active,
but it currently has no effect.

RELNOTES: None.
PiperOrigin-RevId: 193964624