Add execution information to standalone test results.
In StandaloneTestStrategy, copy as much information as SpawnResult makes
available to us through to both the TestResultData and BEP's
TestResult.ExecutionInfo protos. One immediate consequence is that the UI and
BEP can tell you whether a test result was cached remotely.

I changed Executor.getEventHandler to return an ExtendedEventHandler because it
makes this change easier to test.

Change-Id: I94fefdcd2e029c81085076736ad13a4bdf1bae8f