sandbox: Generate actually unique sandbox paths.
Turns out the "unique" id returned by SpawnExecutionContext is only
unique within one SpawnRunner, but not across multiple SpawnRunner
classes.

We have to prefix the generated path with the name of the current
strategy in order to not run into conflicts when a build happens to use
multiple different sandbox strategies.

RELNOTES: None.
PiperOrigin-RevId: 194230475