Automated rollback of commit c5c6ef83936cf90fd11038a84637e691beb47288.
*** Reason for rollback ***

b/78577719

*** Original change description ***

Change action construction to use built-in param file support.

We want to be able to control how and when param files are used, and manual construction of param files prevents this. It should also be less code overall.

For this CL, when param files are used is unchanged, their format and encoding is unchanged. The flag name is also unchanged, except in some cases it was changed from (eg.) "--flagfile foo" to "--flagfile=foo".

However, the name of the param file is now derived from the primary...

***

RELNOTES: None
PiperOrigin-RevId: 194389749