Remove the on-disk caching of Xcode locations.
There is no need for the cache to be on disk.  Originally, there was a
desire to share this cache with other tools... but this never happened.
And, actually, because Bazel is in control of what it runs, it can just
inject the "cached" values into those tools via flags.

Instead, just store the cache in-memory.  This avoids having to open and
read the cache on every single action that is locally executed on a Mac.
Results when building a large iOS app from a clean slate show up to a
1% wall time improvement on my Mac Pro 2013 and a reduction in the
variance of the measurements.

This change also gets rid of the OS check from the action execution's
critical path.  There is not much use in checking this: if we instantiate
this by mistake, the actual calls will fail.  But sometimes we want to
actually run this code on non-macOS systems (e.g. for unit-testing with
mocked tools), so we should allow that.  And this change also ensures
that XcodeLocalEnvProviderTest builds and runs...

RELNOTES: None.
PiperOrigin-RevId: 194681802