Make runfiles usage on Windows more flexible to support remote execution.
When trying to find a runfile on Windows:

1. First look for the runfiles MANIFEST and find runfile locations using this if it exists (current behavior).

2. If no MANIFEST file exists, look for runfiles in the runfiles directory (new behavior).

As part of this, remove setting RUNFILES_MANIFEST_ONLY for the benefit of test-setup.sh. Instead of telling it what to do, it decides what to do based on the observed state of the world. Launchers still set RUNFILES_MANIFEST_ONLY for the benefit of launched programs, since some may depend on this.

Fixes https://github.com/bazelbuild/bazel/issues/4962.

RELNOTES: Remote execution works for Windows binaries with launchers.
PiperOrigin-RevId: 194785440