remote/http: fully read http content on error status
When the http response has a status other than 200 and
also has a Content-Length header set, then wait until
all content has been received before completing the user
promise.

In case of any errors, close the channel in order to make
sure it's not reused as we don't know what data is left
on the wire.

Closes #5101.

PiperOrigin-RevId: 194787393