Fix CppConfiguration#isLLVMOptimizedFdo() method to work with --fdo_profile
This method is used to determine whether a zipper executable should be loaded. Currently it only checks whether the file specified by --fdo_optimize is a zip file. As we cannot check whether --fdo_profile points to a zip file in this phase, we load it unconditionally in the case of --fdo_profile optimization.

PiperOrigin-RevId: 195087124